requirements.txt
transformers>=4.36.2
sentence-transformers>=2.2.2
pypdf>=3.17.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.23.2
huggingface-hub>=0.20.3
tqdm>=4.66.1
numpy>=1.26.2
